requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.0.2
asyncpg==0.29.0
pydantic==2.5.0
python-multipart==0.0.6
pytesseract==0.3.10
spacy==3.7.2
pillow==10.1.0
python-dotenv==1.0.0
httpx>=0.24.0,<0.25.0